debugging

+++ Code/Models/generator_training_class.py
@@ -150,7 +150,7 @@ class generator:
                 # trim input, target
                 input = torch.from_numpy(
                     input[:seq_length_input.max()]
-                    ).float().to(self.device)
+                    ).long().to(self.device)
                 target = torch.from_numpy(
                     target[:seq_length_target.max()]
                     ).long().to(self.device)
